+++ odbc_tests/tests/odbc-api/catalog/table_privileges_tests.cpp
@@ -33,7 +33,7 @@ TEST_CASE_METHOD(ReadOnlyDbStmtFixture, "SQLTablePrivileges: Result set has corr
   REQUIRE(numCols == 7);
 }
 
-TEST_CASE_METHOD(ReadOnlyDbStmtFixture, "SQLTablePrivileges: Result set column names match ODBC 3.x spec",
+TEST_CASE_METHOD(ReadOnlyDbStmtFixture, "SQLTablePrivileges: Result set names and string types match ODBC 3.x spec",
                  "[odbc-api][catalog][tableprivileges]") {
   SQLRETURN ret = SQLTablePrivileges(stmt_handle(), sqlchar(database_name()), SQL_NTS, sqlchar(schema_name()), SQL_NTS,
                                      sqlchar(readonly_db::BASIC_TABLE), SQL_NTS);
@@ -54,6 +54,7 @@ TEST_CASE_METHOD(ReadOnlyDbStmtFixture, "SQLTablePrivileges: Result set column n
                          &colSize, &decDigits, &nullable);
     REQUIRE(ret == SQL_SUCCESS);
     REQUIRE(std::string(colName) == expectedColNames[col - 1]);
+    REQUIRE(dataType == SQL_WVARCHAR);
   }
 }